refactor: Load configuration via json file
This change is made to be able to work with the same code files on the AWS and the local machine. 
The configuration consists by now of the paths to the car-sim-video-images and the results that might be checkpoints, and/or videos of the autonomous driving.
Later parameters like epochs and batch:size can be added, to account for different computational rescources on different machines.