feat: Build residuum convolutional block in keras
Building the residuum block was just a prove of concept to me, that these type of more complex models can also be build with the keras API. 
The main difference to the usual way of defining a model in keras is, that one has to define an input and and output tensor with some desired number of layers in between. The model is then build at the end with 'model = Model(init, out, name='myCNN')'.