fix: Access pd DataFrame via iloc after shuffling
The file 'driving_log.csv' is read with pandas and saved as 'dataset'. 
In order the get random batches, the dataset gets shuffled with each batch. However, if one accesses it then with
'dataset[some_key][idx]' 
one will always get the same value.
The solution is 
'dataset[some_key].iloc[idx]' 
which will return the idxth value of the shuffled dataset.