feat: Add affine transformation to pipeline
The code is adapted from:
    https://medium.com/@ksakmann/behavioral-cloning-make-a-car-drive-like-yourself-dc6021152713#.7k8vfppvk
The added function performs an affine transformation on the original image. This results in a new image, which still has a valid perspective but with a significantly altered steering angle.
This allows to generate any real valued steering angle within the valid range defined by the shearing range.
The main result is a smooth distributed dataset w.r.t. the steering angle.