fix: Change arguments for fit_generator
The argument to the fit_generator were wrong. I missunderstop the notion, that one no longer needs to specifiy the total size of the training set but instead only defines the batch size, the number of epochs and the steps per epoch.
The same applies for the validation data set arguments.